refactor: move tokens.js to internal
