+++ packages/vats/src/core/startWalletFactory.js
@@ -6,7 +6,7 @@ import { makeTracer, VBankAccount } from '@agoric/internal';
 import { AmountMath } from '@agoric/ertp';
 import { ParamTypes } from '@agoric/governance';
 import { makeStorageNodeChild } from '@agoric/internal/src/lib-chainStorage.js';
-import { Stable } from '@agoric/inter-protocol/src/tokens.js';
+import { Stable } from '@agoric/internal/src/tokens.js';
 import {
   makeHistoryReviver,
   makeBoardRemote,
