+++ packages/vats/test/test-demoAMM.js
@@ -1,6 +1,6 @@
 import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
 import { AmountMath, makeIssuerKit } from '@agoric/ertp';
-import { Stable } from '@agoric/inter-protocol/src/tokens.js';
+import { Stable } from '@agoric/internal/src/tokens.js';
 
 import {
   AMMDemoState,
